feat: layout.tsx を Server Component 化し Header.tsx を切り出す（タスク1-1）
"use client" を削除し metadata API を有効化。ナビゲーション・SPドロワーの
インタラクション（useState）は Header.tsx（Client Component）に移譲。
